LayoutTests/sputnik/Conformance/12_Statement/12.6_Iteration_Statements/12.6.3_The_for_Statement/S12.6.3_A8_T1-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Unexpected token '++'. Expected a ':' following the property name 'index'.
S12.6.3_A8_T1

PASS Expected parsing failure

TEST COMPLETE
